ImagesListPresenter: добавлен ининциализатор для удобства; ProfileViewController: выненсена логика logout в ProfilePresenter
